fix(fse): close DoS overflow + tighten Predefined cache copy
#20: probability sum overflow in u32. Validate per-probability range
(must be in {-1, 0, 1..=table_size}) upfront and compute the sum in
u64 so the wrap-around exploit (large positive i32 cast to u32 then
summed wrapping back to table_size) is impossible. Bump
FSETableError::ProbabilityCounterMismatch.got to u64 to match the
widened sum type — wire-format parse path is unaffected (uses u32
internally then upcasts at the error site).

#21/#22/#23: predefined LL/ML/OF cache hot-path used clone_from on
FSETable, which also clones symbol_spread_buffer + symbol_probabilities
(both build-only fields not touched on the decode path). Switch to
reinit_from, which skips symbol_spread_buffer content copy (reserves
capacity only) and resets the buffer to empty. The decode Vec stays
fully copied via extend_from_slice so the enriched Entry fields
(base_value, num_additional_bits) carry through unchanged.

+++ zstd/src/decoding/errors.rs
@@ -1122,7 +1122,7 @@ pub enum FSETableError {
     },
     GetBitsError(GetBitsError),
     ProbabilityCounterMismatch {
-        got: u32,
+        got: u64,
         expected_sum: u32,
         symbol_probabilities: Vec<i32>,
     },
